build: built a video streaming server
During the weekend(28th November, 2025), I decided to bring out some
free time and learn how to send media data(i.e. music audio, videos)
through a server in a network. I learnt about streams and how they work
to ensure that even big files can be sent to a device across the
Internet.

The next thing for me is to learn how to transcode data like Youtube. If
someone wants to select a higher or lower video quality(e.g. 144p, 720p)
, he can do that.

So yeah.
That's it for now.

Till the next commit 👋